Style: split components
